feat(driver): add dds texture support at packer
